build: Drop PHP 5.5 support and limit HHVM support to 3.18 (#27)
This limits support to PHP versions that are currently supported
or were supported in the last 6 months.

Also remove HHVM versions that aren't supported by Wikimedia
from test matrix to make tests faster and because issues in them
would not be prioritised. The next release will likely drop support
entirely.